set gradle to 4.1.0

+++ dlnacontroller/src/main/cpp/MediaRemote_jni.cpp
@@ -114,7 +114,7 @@ static MediaRemote* setMediaRemote(JNIEnv* env, jobject thiz, MediaRemote* conte
     if(gMediaRemoteObjectLock)
         LockMutex(gMediaRemoteObjectLock);
     MediaRemote* old = (MediaRemote*)env->GetIntField(thiz, gMediaRemoteClassFields.mNativeContext);
-    env->SetIntField(thiz, gMediaRemoteClassFields.mNativeContext, (int)context);
+    env->SetIntField(thiz, gMediaRemoteClassFields.mNativeContext, (long)context);
     if(gMediaRemoteObjectLock)
         UnlockMutex(gMediaRemoteObjectLock);
     return old;
